requirements.txt
# Core Django
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1

# API Documentation
drf-yasg==1.21.7

# Database
mysqlclient==2.2.0
# psycopg2-binary==2.9.9  # Optional: PostgreSQL support (install if needed)

# Background Task Scheduling
django-apscheduler==0.6.2

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
flake8==6.1.0
isort==5.13.2
pylint==3.0.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
